+++ packages/apollo-language-server/src/engine/index.ts
@@ -46,7 +46,7 @@ export type FieldStats = Map<string, Map<string, number | null>>;
 export function noServiceError(service: string | undefined, endpoint?: string) {
   return `Could not find service ${
     service ? service : ""
-  } from Engine at ${endpoint}. Please check your API key and service name`;
+  } from Apollo Graph Manager at ${endpoint}. Please check your API key and service name`;
 }
 
 export class ApolloEngineClient extends GraphQLDataSource {
@@ -98,7 +98,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service;
     });
@@ -121,7 +121,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service.checkSchema;
     });
@@ -144,7 +144,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service.uploadSchema;
     });
@@ -169,7 +169,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service.upsertImplementingServiceAndTriggerComposition;
     });
@@ -194,7 +194,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service.checkPartialSchema;
     });
@@ -212,7 +212,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!data || !data.service) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
 
       return data.service.removeImplementingServiceAndTriggerComposition;
@@ -236,7 +236,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       }
 
       if (!(data && data.service)) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
 
       return data.service.validateOperations.validationResults;
@@ -262,7 +262,7 @@ export class ApolloEngineClient extends GraphQLDataSource {
       if (
         !(data && data.service && data.service.registerOperationsWithResponse)
       ) {
-        throw new Error("Error in request from Engine");
+        throw new Error("Error in request from Apollo Graph Manager");
       }
       return data.service.registerOperationsWithResponse;
     });