+++ packages/apollo/src/commands/service/push.ts
@@ -9,7 +9,7 @@ import chalk from "chalk";
 
 export default class ServicePush extends ProjectCommand {
   static aliases = ["schema:publish"];
-  static description = "Push a service to Engine";
+  static description = "Push a service to Apollo Graph Manager";
   static flags = {
     ...ProjectCommand.flags,
     tag: flags.string({
@@ -48,10 +48,10 @@ export default class ServicePush extends ProjectCommand {
     let gitContext;
     await this.runTasks(({ flags, project, config }) => [
       {
-        title: "Uploading service to Engine",
+        title: "Uploading service to Apollo Graph Manager",
         task: async () => {
           if (!config.name) {
-            throw new Error("No service found to link to Engine");
+            throw new Error("No service found to link to Apollo Graph Manager");
           }
 
           if (flags.federated) {
@@ -130,9 +130,9 @@ export default class ServicePush extends ProjectCommand {
           };
 
           const { schema: _, ...restVariables } = variables;
-          this.debug("Variables sent to Engine:");
+          this.debug("Variables sent to Apollo Graph Manager:");
           this.debug(restVariables);
-          this.debug("SDL of introspection sent to Engine:");
+          this.debug("SDL of introspection sent to Apollo Graph Manager:");
           this.debug(printSchema(schema));
 
           const response = await project.engine.uploadSchema(variables);
@@ -143,7 +143,7 @@ export default class ServicePush extends ProjectCommand {
               hash: response.tag ? response.tag.schema.hash : null,
               code: response.code
             };
-            this.debug("Result received from Engine:");
+            this.debug("Result received from Apollo Graph Manager:");
             this.debug(result);
           }
         }